fix: Refactor probe flags
The probe flags are defined in `cmd/probe` and then used in `cmd/runner` -- this coupling is implicit and error prone.

This refactor pulls the probe flag strings out into a common package and defines a function to turn the strings into parseable flags. You can execute tests for this with `make test`.