Pass PROBE_VERSION as version instead of full -X flag
We were getting a build error with the previous version, which was
pointing at the `-X` flag of `go build -ldflags` not being properly
formatted. By only passing the version instead of the whole import
path for overriding the value at compile time we hope it will fix this
error.

    #10 9.570 /usr/local/go/pkg/tool/linux_amd64/link: -X flag requires argument of the form importpath.name=value