Refactor `pkg/kubernetes` (#57)
This PR refactos pkg/kubernetes to make it a bit more Goish™, and also tries add some missing unit tests. I think the full files view should be enough, but in any case each commit is trying to be as small and scoped as possible so it might be easier to review that way.

* Refactor Kubernetes.GetPod

This function wasn't being used, only tested. Here we refactor it so
we can eventually encapsulate this logic in this struct, and add
proper testing.

* Use Kubernetes.GetPods for finding pods

* Hide low-level Kubernetes fields

These aren't necessary for proper utilization of this struct.

* Rename kubernetes.GetKubernetes to kubernetes.New

* Refactor kubernetes.New

* Remove unused field

* Refactor chooseTargetContainer

* Add getEphemeralContainerExitCode unit test

* Add isEphemeralContainerTerminated unit test

* Revert get cluster config logic

See [note] on pull request, we should probably fix this.

[note]: https://github.com/grafana/nethax/pull/57#issuecomment-2890792252