Return error if ephemeral container is not found
When checking the ephemeral container status, we were previously
ignoring the fact that the ephemeral container doesn't existed, and
just reporting that it hasn't terminated. This was hiding the fact
that the ephemeral container *should be present*, and that if not,
then it could be indicating that there was an error somewhere else.