Refactor polling for ephemeral container terminated state
This consolidates the three different functions we had for polling the
terminated state of the ephemeral container into a single function, to
avoid duplicating the code and reducing the number of lines to
maintain. It also adds a, probably too clever, test setup that test
the different stages the ephemeral container goes through during its
lifetime.