Change Renovate schedule for grafana/sharedworkflows (#143)
This dependency gets updated very frequently, triggering Renovate to
create a PR almost on a daily basis, but the action that we use, at
the time of writing this, hasn't been updated in 2 months. I couldn't
use the tag for this action as Zizmor complains that we are not using
a hash, and we would like to keep updating this eventually, so let's
change the frequency at which Renovate updates this action to once a
week instead.