Consolidate actions/checkout version
In both #181 and #182 Renovate is trying to update the
`actions/checkout` version to the latest, but in both cases it is
using a different hash. This PR consolidates both places where we use
this action with the latest value (v6.0.0) found in the repository.

More info at https://github.com/actions/checkout

Deprecates #181
Deprecates #182